backend/requirements.txt

fastapi==0.116.1
joblib==1.5.2
numpy==1.26.4
pandas==2.3.3
pip==25.2
psycopg2-binary==2.9.11
pydantic==2.11.9
requests==2.32.5
scikit-learn==1.2.2
SQLAlchemy==2.0.44
tqdm==4.67.1
uvicorn==0.35.0
xgboost==2.0.3
fastapi-cors
python-dotenv